Order History (Backend only)

+++ server/RevShop/src/main/java/net/revature/RevShop/Services/UserService.java
@@ -27,7 +27,7 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public Optional<User> getUserById(long userId) {
+    public Optional<User> getUserById(Integer userId) {
 
         return userRepository.findById(userId);
     }
@@ -45,7 +45,7 @@ public class UserService {
         return userRepository.existsByUsername(username);
     }
 
-    public User updateUser(Long userId, User updatedUser) throws IllegalArgumentException {
+    public User updateUser(Integer userId, User updatedUser) throws IllegalArgumentException {
         User existingUser = userRepository.findById(userId).orElseThrow(() -> new IllegalArgumentException("User not found"));
 
         if (!existingUser.getUsername().equals(updatedUser.getUsername()) &&
@@ -68,7 +68,7 @@ public class UserService {
         return userRepository.save(existingUser);
     }
 
-    public void deleteUser(Long userId) {
+    public void deleteUser(Integer userId) {
         userRepository.deleteById(userId);
     }
